Refactoring for vertical architecture: Role slice

+++ src/Modules/HRManagement.Modules.Staff/Data/Configurations/EmployeeConfiguration.cs
@@ -48,7 +48,7 @@ public class EmployeeConfiguration : IEntityTypeConfiguration<Employee>
                 })
             .Navigation(x => x.TerminationDate);
         builder.HasOne(x => x.Role).WithMany().IsRequired();
-        builder.Property<byte>("RoleId").HasColumnName("RoleId").IsRequired();
+        builder.Property<int>("RoleId").HasColumnName("RoleId").IsRequired();
         builder.HasOne(x => x.Manager).WithMany();
         builder.Property<Guid?>("ManagerId").HasColumnName("ManagerId");
         builder.HasMany(x => x.ManagedEmployees).WithOne(x => x.Manager);
