+++ src/Modules/HRManagement.Modules.Staff/Features/GetRoles/RoleDto.cs
@@ -4,7 +4,7 @@ namespace HRManagement.Modules.Staff.Features.FindRoles;
 
 public class RoleDto
 {
-    [JsonPropertyName("id")] public byte Id { get; set; }
+    [JsonPropertyName("id")] public int Id { get; set; }
     [JsonPropertyName("name")] public string Name { get; set; }
     [JsonPropertyName("manager_role")] public string ReportsTo { get; set; }
 }
